chore: update routing and configuration for improved SEO and user experience
- Refactored middleware to handle redirection from root to the homepage, enhancing navigation consistency.
- Updated Next.js configuration to include a redirects function for better routing management.
- Enhanced robots.txt to allow indexing of the root page while disallowing app-related paths.
- Improved sitemap generation to reflect the new routing structure.
- Removed deprecated patterns from ESLint configuration for cleaner code.
- Updated NotFound and Error components to provide user-friendly messages and links to the homepage.
- Added loading and error handling components to enhance user experience during data fetching.
