+++ src/app/(app)/error.tsx
@@ -3,7 +3,7 @@
 import { Button } from "@/components/ui/button";
 import { useEffect } from "react";
 import { toast } from "sonner";
-import { INITIAL_INVOICE_DATA } from "../../constants";
+import { INITIAL_INVOICE_DATA } from "../constants";
 import { umamiTrackEvent } from "@/lib/umami-analytics-track-event";
 import * as Sentry from "@sentry/nextjs";
 import { ErrorMessage } from "@/components/etc/error-message";
