+++ src/app/robots.ts
@@ -8,26 +8,17 @@ export default function robots(): MetadataRoute.Robots {
       {
         userAgent: "*",
         allow: [
+          // Allow main app page (now at /)
+          "/",
           // Allow about pages in all languages
           ...SUPPORTED_LANGUAGES.map((locale) => `/${locale}/about`),
-          // Allow main app page (/en/app) for now
-          "/en/app",
         ],
         disallow: [
-          // Disallow root as it redirects to /en/app
-          "/",
-          // Disallow shared invoice URLs, like /en/app?data=*
-          "/*app?data=*",
-          "/*app?*data=*",
+          // Disallow shared invoice URLs, like /?data=*
+          "/?data=*",
           // Disallow subscription confirmation pages with and without tokens
           "/confirm-subscription",
           "/confirm-subscription?*",
-          ...SUPPORTED_LANGUAGES.flatMap((locale) => [
-            `/${locale}/confirm-subscription`,
-            `/${locale}/confirm-subscription?*`,
-          ]),
-          // Disallow any other query parameters
-          "/*?*",
         ],
       },
     ],
